Sample html
Note again that the JS may be off. 
Also note - this code does not work for percents > 50%. For that, you have to switch the colors or reversing the sign might do it too. I didn't make it dynamic because this was for a marketing thing and they don't talk about percentages less than 50%.